Remove `CACHE_KEY` computation logic that never run
We can say with confidence that it never run because 1) all the callers
in this plugin pass a cache and 2) because of the `u` flag in the
shebang, the line `CACHE_KEY=$2` would have failed the script if no
value had been passed, therefore never running the fallback logic.
Moreover, the script uses `$3` and `$4` later on, further proving that
`$2` was always, implicitly, required.