Add `pnginfo` utility to check PNG files
Mostly useful to check if a PNG file is of expected size and if it has an alpha channel or not. Typical use case for this is to have a CI check validate that iOS App Icons are 1024x1024 and don't have an alpha channel (otherwise they'd be rejected from review), or to validate the size of App Store screenshots.