Clean up Makefile
We had a mix of using docker-compose and using docker run that made the Makefile a bit inconsistent.

I decided to adopt the same approach as we did in `buildkite-ci` and get rid of the `docker-compose.yml` file and only use the `Makefile` to call `docker run`, to avoid any extra unnecessary indirections when trying to make sense of what gets run.

PS: Besides, using `docker-compose` for `shellcheck hooks/** bin/**` wouldn't have been possible anyway, because referencing `bin/**` in `docker-compose.yml` would pass the literal 'bin/**' as an argument without shell-expanding it to the list of files in `bin/`, which would make `shellcheck` error with a 'file not found'.