[AndroidManifest] Improve Merged Manifest Diff (#173)
* Android: Merge base branch changes into head branch prior mm generation

As reported already, on some PRs, version code/name changes are detected
because the head branch wasn't synced with the base branch. In order to
avoid such false positives, this change merges the base branch into
the head prior to generating the head manifest. This way, comparing the
base and head manifest changes will always produce the most appropriate
result.

* Release: Add changelog entry for merged manifest diff improvement